Move FindSkillsModal into components folder
